feat(JA): Code refactor and moving functions around for code sharing with upcoming JA CLI.
