+++ src/deadline/client/api/_submit_job_bundle.py
@@ -13,7 +13,9 @@ import textwrap
 from configparser import ConfigParser
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
-from botocore.client import BaseClient  # type: ignore[import]
+from botocore.client import BaseClient
+
+from deadline.client.api._job_attachment import _hash_attachments  # type: ignore[import]
 
 from .. import api
 from ..exceptions import DeadlineOperationError, CreateJobWaiterCanceled
@@ -35,12 +37,11 @@ from ..job_bundle.submission import AssetReferences, split_parameter_args
 from ...job_attachments.exceptions import MisconfiguredInputsError
 from ...job_attachments.models import (
     JobAttachmentsFileSystem,
-    AssetRootGroup,
     AssetRootManifest,
     AssetUploadGroup,
     JobAttachmentS3Settings,
 )
-from ...job_attachments.progress_tracker import SummaryStatistics, ProgressReportMetadata
+from ...job_attachments.progress_tracker import ProgressReportMetadata
 from ...job_attachments.upload import S3AssetManager
 from ._session import session_context
 
@@ -401,42 +402,6 @@ def wait_for_create_job_to_complete(
     raise TimeoutError(
         f"Timed out after {delay_sec * max_attempts} seconds while waiting for Job to be created: {job_id}"
     )
-
-
-def _hash_attachments(
-    asset_manager: S3AssetManager,
-    asset_groups: list[AssetRootGroup],
-    total_input_files: int,
-    total_input_bytes: int,
-    print_function_callback: Callable = lambda msg: None,
-    hashing_progress_callback: Optional[Callable] = None,
-    config: Optional[ConfigParser] = None,
-) -> Tuple[SummaryStatistics, List[AssetRootManifest]]:
-    """
-    Starts the job attachments hashing and handles the progress reporting
-    callback. Returns a list of the asset manifests of the hashed files.
-    """
-
-    def _default_update_hash_progress(hashing_metadata: Dict[str, str]) -> bool:
-        return True
-
-    if not hashing_progress_callback:
-        hashing_progress_callback = _default_update_hash_progress
-
-    hashing_summary, manifests = asset_manager.hash_assets_and_create_manifest(
-        asset_groups=asset_groups,
-        total_input_files=total_input_files,
-        total_input_bytes=total_input_bytes,
-        hash_cache_dir=config_file.get_cache_directory(),
-        on_preparing_to_submit=hashing_progress_callback,
-    )
-    api.get_deadline_cloud_library_telemetry_client(config=config).record_hashing_summary(
-        hashing_summary
-    )
-    print_function_callback("Hashing Summary:")
-    print_function_callback(textwrap.indent(str(hashing_summary), "    "))
-
-    return hashing_summary, manifests
 
 
 @api.record_success_fail_telemetry_event(metric_name="cli_asset_upload")  # type: ignore
